+++ docs/changelog.md
@@ -63,6 +63,10 @@
 
 ## 更新说明详情
 
+1.11.14 fix: fix migration_area_baidu interface
+
+    1. 修复 migration_area_baidu 接口
+
 1.11.13 fix: fix futures_comm_info interface
 
     1. 修复 futures_comm_info 接口
@@ -2882,6 +2886,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.11.14 fix: fix migration_area_baidu interface
 
 1.11.13 fix: fix futures_comm_info interface
